Broke out preload functions into ooi_funcs module and global vars into ooi_gobals modukle
