+++ ooi_globals.py
@@ -1,0 +1,16 @@
+# M2M URLs
+BASE_URL = 'https://ooinet.oceanobservatories.org/api/m2m/'
+PRELOAD_URL = BASE_URL + '12575/stream/byname/'
+U = BASE_URL + '/12576/sensor/inv/'
+
+# Verbose? (boolean)
+VERB = True
+
+# Engineering Only or Science Only? (True=Eng,False=Sci)
+ENG_ONLY = False
+
+# Setup Nodes of Interest List
+wanted_nodes = ['LJ01A', 'MJ01A', 'SF01A', 'PC01A', 'LJ01B', 'DP01A',
+                'MJ01B', 'LJ01C', 'SF01B', 'PC01B', 'DP01B', 'MJ01C',
+                'LJ01D', 'MJ03A', 'LJ03A', 'PC03A', 'SF03A', 'MJ03B',
+                'MJ03E', 'MJ03F', 'DP03A', 'MJ03C', 'MJ03D']
